# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A production-ready RESTful API for managing movies with comprehensive user authentication, role-based authorization, rate limiting, and email notifications.\n\n## Features\n- Full CRUD operations for movies\n- User registration and authentication\n- Role-based access control (RBAC)\n- Token-based authentication (Bearer)\n- Email verification and password reset\n- Rate limiting (2 req/s, burst: 4)\n- Pagination and filtering\n- CORS support\n\n## Authentication\nMost endpoints require authentication. Use the `/v1/tokens/authentication` endpoint to obtain a token, then include it in the Authorization header as: `Bearer YOUR_TOKEN`\n\n## Rate Limiting\nAPI requests are rate-limited to 2 requests per second with a burst of 4 requests.",
        "title": "Gmoapi - Movie Management API",
        "contact": {
            "name": "API Support Team",
            "url": "https://github.com/ucok-man/gmoapi",
            "email": "support@gmoapi.ucokman.web.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "paths": {
...
        "/movies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of movies with optional filtering by title (full-text search) and genres. Supports sorting by multiple fields.\n\n**Permissions Required:** `movies:read`\n\n**Filtering:**\n- Title: Partial match using PostgreSQL full-text search\n- Genres: Multiple genres can be specified (comma-separated)\n\n**Sorting:**\n- Prefix with `-` for descending order (e.g., `-year`)\n- Available fields: id, title, year, runtime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "List All Movies",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Godfather",
                        "description": "Filter by movie title (partial match, case-insensitive)",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "drama,crime",
                        "description": "Filter by genres (comma-separated)",
                        "name": "genres",
                        "in": "query"
                    },
                    {
                        "maximum": 10000000,
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (minimum: 1, maximum: 10,000,000)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page (minimum: 1, maximum: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "title",
                            "year",
                            "runtime",
                            "-id",
                            "-title",
                            "-year",
                            "-runtime"
                        ],
                        "type": "string",
                        "default": "id",
                        "description": "Sort field",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of movies with pagination metadata",
                        "schema": {
                            "type": "object",
                            "properties": {
                                " metadata": {
                                    "$ref": "#/definitions/data.Metadata"
                                },
                                "movies": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/data.Movie"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid query parameters",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid authentication token",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user account not activated or insufficient permissions",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - validation errors",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests - rate limit exceeded",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new movie entry in the catalog. All fields are required.\n\n**Permissions Required:** `movies:write`\n\n**Validation Rules:**\n- Title: Required, max 500 characters\n- Year: Required, between 1888 and current year\n- Runtime: Required, positive integer, format \"123 mins\"\n- Genres: Required, 1-5 unique genres",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Create New Movie",
                "parameters": [
                    {
                        "description": "Movie creation data",
                        "name": "movie",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                " genres": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                " runtime": {
                                    "type": "string"
                                },
                                " year": {
                                    "type": "integer",
                                    "format": "int32"
                                },
                                "title": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Movie created successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "movie": {
                                    "$ref": "#/definitions/data.Movie"
                                }
                            }
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "URL of the created movie"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - malformed JSON or invalid data types",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid authentication token",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user account not activated or insufficient permissions",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity - validation errors",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests - rate limit exceeded",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete a movie from the catalog by its ID. This action cannot be undone.\n\n**Permissions Required:** `movies:write`",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Delete Movie",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "Movie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie deleted successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid authentication token",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - user account not activated or insufficient permissions",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests - rate limit exceeded",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "data.Metadata": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "first_page": {
                    "type": "integer"
                },
                "last_page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
        "data.Movie": {
            "type": "object",
            "properties": {
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "runtime": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "version": {
                    "description": "The version number starts at 1 and will be incremented\neach time the movie information is updated",
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}